feat: add search validation
Using the search input just to search for a new income/outcome, without checking whether any content appears on the screen results in a poor quality test.

This commit adds a search validation by getting the text of the income/outcome and validating if it is the same text that was used on the search input.